+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 UNRELEASED
 
+- Implement SearchFilesBy{Content,Name}
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/677
 - Introduce feature flag package based on gRPC metadata
   https://gitlab.com/gitlab-org/gitaly/merge_requests/704